# fix(aws): reducing timeouts to speed up error feedback loop