--- a/pkg/aws/aws.go
+++ b/pkg/aws/aws.go
@@ -88,11 +88,11 @@ const (
 	maxResults int64 = 50
 	pluginName       = "session-manager-plugin"
 	// API calls retry configuration
-	numMaxRetries    = 10
-	minRetryDelay    = 1 * time.Second
-	minThrottleDelay = 2 * time.Second
-	maxRetryDelay    = client.DefaultRetryerMaxRetryDelay
-	maxThrottleDelay = client.DefaultRetryerMaxThrottleDelay
+	numMaxRetries    = 5
+	minRetryDelay    = 10 * time.Millisecond
+	minThrottleDelay = 500 * time.Millisecond
+	maxRetryDelay    = 1 * time.Second
+	maxThrottleDelay = 4 * time.Second
 	// TargetTypeInstanceID points to an instance ID type
 	TargetTypeInstanceID = "instance-id"
 	// TargetTypePrivateDNS points to a private DNS type