I've noticed extremely slow error feedback loop e.g. for broken credentials. I decided to keep a feedback loop significantly faster for non-throttled errors.

Benchmarked with an expired token errors (non-throttled):

```console
$ hyperfine -i 'bazel-out/darwin-fastbuild-ST-cff33a8e9d06/bin/dev/sigil --region eu-central-1 --log-level fatal ls'
Benchmark #1: bazel-out/darwin-fastbuild-ST-cff33a8e9d06/bin/dev/sigil --region eu-central-1 --log-level fatal ls
  Time (mean ± σ):      1.621 s ±  2.044 s    [User: 13.7 ms, System: 41.3 ms]
  Range (min … max):    0.852 s …  7.436 s    10 runs
```

For throttled environments it should still capable to work without errors (CC: @klauern) even though times and number of retries are reduced.